Improve api to use same resource to handle post and delete
