+++ components/org.wso2.micro.integrator.extensions/org.wso2.micro.integrator.management.apis/src/main/java/org/wso2/micro/integrator/management/apis/Constants.java
@@ -25,6 +25,7 @@ public class Constants {
     public static final String REST_API_CONTEXT = "/management";
     public static final String PREFIX_APIS = "/apis";
     public static final String PREFIX_CARBON_APPS = "/applications";
+    public static final String PATH_PARAM_CARBON_APP_NAME = "/" + "{name}";
     public static final String PREFIX_ENDPOINTS = "/endpoints";
     public static final String PREFIX_INBOUND_ENDPOINTS = "/inbound-endpoints";
     public static final String PREFIX_PROXY_SERVICES = "/proxy-services";
@@ -37,8 +38,6 @@ public class Constants {
     public static final String PREFIX_TEMPLATES = "/templates";
     public static final String PREFIX_MESSAGE_STORE = "/message-stores";
     public static final String PREFIX_MESSAGE_PROCESSORS = "/message-processors";
-    public static final String PREFIX_CAPP_DEPLOYER = "/capp-deployer";
-    public static final String PREFIX_CAPP_UNDEPLOYER = "/capp-undeployer";
     public static final String PREFIX_EXTERNAL_VAULTS = "/external-vaults";
     public static final String PATH_PARAM_EXTERNAL_VAULT_NAME = "/" + "{vault}";
     public static final String PREFIX_LOCAL_ENTRIES = "/local-entries";
